Rip out the method for SearchOptions from ElasticsearchService
